+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 0.14.2
 
+### Enhancements
+- Queries can now be constructed by chosing concepts in another data source that are linked to the original query data source
+
 ### Refactoring
 - Removed uri, namespace, and name from concepts for simpler imports and exports
 - Data Dictionary import simplified, imports CSV data by column header name